+++ src/main/webapp/components/BaseFormView.jsx
@@ -888,7 +888,7 @@ class BaseFormView extends PureComponent {
                                         serviceName={this.props.serviceName}
                                         mode={this.props.mode}
                                         disabled={temState.disabled}
-                                        markDownMessage={temState.markDownMessage}
+                                        markdownMessage={temState.markdownMessage}
                                         dependencyValues={temState.dependencyValues || null}
                                     />
                                 );
@@ -958,7 +958,7 @@ class BaseFormView extends PureComponent {
                                 serviceName={this.props.serviceName}
                                 mode={this.props.mode}
                                 disabled={temState.disabled}
-                                markDownMessage={temState.markDownMessage}
+                                markdownMessage={temState.markdownMessage}
                                 dependencyValues={temState.dependencyValues || null}
                             />
                         );
