+++ src/main/webapp/components/ControlWrapper.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import ControlGroup from '@splunk/react-ui/ControlGroup';
 import styled from 'styled-components';
 
-import MarkDownMessage from './MarkDownMessage';
+import MarkdownMessage from './MarkdownMessage';
 import CONTROL_TYPE_MAP from '../constants/ControlTypeMap';
 
 const CustomElement = styled.div`
@@ -44,7 +44,7 @@ class ControlWrapper extends React.PureComponent {
         const { handleChange, addCustomValidator, utilCustomFunctions } = this.props.utilityFuncts;
         // We have to put empty object because markDownMessage prop can be undefined 
         // because we are not explicitly setting it but expecting it from custom hooks only.
-        const { text, link, color, markdownType, token, linkText } = this.props.markDownMessage || {};
+        const { text, link, color, markdownType, token, linkText } = this.props.markdownMessage || {};
         let rowView;
         if (this.props.entity.type === 'custom') {
             const data = {
@@ -80,7 +80,7 @@ class ControlWrapper extends React.PureComponent {
 
         let helpText = (
             <>
-                <MarkDownMessage
+                <MarkdownMessage
                     text={ text || '' }
                     link={ link || '' }
                     color={ color || ''}
@@ -116,7 +116,7 @@ ControlWrapper.propTypes = {
     error: PropTypes.bool,
     entity: PropTypes.object,
     disabled: PropTypes.bool,
-    markDownMessage: PropTypes.object,
+    markdownMessage: PropTypes.object,
     serviceName: PropTypes.string,
     dependencyValues: PropTypes.object,
 };
